refactor: split oversized cli and ipcctrl files along their concerns
internal/cli/root.go was 717 lines of which 72% were the ten
Build*Command cobra factories — generic command builders that every
generated CLI command goes through, buried under the process
bootstrap. They move to command_builders.go; root.go keeps only the
root command, Execute, init and daemon-dispatch helpers (191 lines).

internal/app/ipcctrl/info.go was 726 lines mixing three concerns. It
splits along the seams the package's actions_*.go files already
follow: info.go keeps the handler type, status and config reads;
info_config.go takes the four config-mutation handlers; info_health.go
takes the health handler and the capability/profile serialization that
backs neru doctor.

Pure file moves — no function bodies changed.
