+++ internal/app/ipcctrl/info.go
@@ -2,10 +2,8 @@ package ipcctrl
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"sync"
 
 	"go.uber.org/zap"
@@ -312,415 +310,4 @@ func (h *InfoHandler) handleReloadConfig(ctx context.Context, _ ipc.Command) ipc
 		Message: "config reloaded successfully",
 		Code:    ipc.CodeOK,
 	}
-}
-
-func (h *InfoHandler) handleHealth(ctx context.Context, _ ipc.Command) ipc.Response {
-	hasErrors := false
-	// --- component checks ---------------------------------------------------
-	components := make(map[string]string)
-	capabilities := capabilitiesMap(h.systemCapabilities())
-	// Event tap — only enabled during active modes (hints/grid/scroll),
-	// so "disabled" in idle mode is expected and healthy.
-	if h.eventTap != nil {
-		if h.eventTap.IsEnabled() {
-			components["event_tap"] = "ok (active)"
-		} else {
-			components["event_tap"] = "ok (idle)"
-		}
-	} else {
-		components["event_tap"] = healthNotInitialized
-		hasErrors = true
-	}
-	// IPC server (implicitly healthy since we're responding, but verify)
-	if h.ipcServer != nil {
-		if h.ipcServer.IsRunning() {
-			components["ipc_server"] = "ok"
-		} else {
-			components["ipc_server"] = "not running"
-			hasErrors = true
-		}
-	} else {
-		components["ipc_server"] = healthNotInitialized
-		hasErrors = true
-	}
-
-	cfg := h.configSnapshot()
-	if cfg != nil {
-		validateErr := cfg.Validate()
-		if validateErr != nil {
-			components["config"] = validateErr.Error()
-			hasErrors = true
-		} else {
-			components["config"] = "ok"
-		}
-	} else {
-		components["config"] = "not loaded"
-		hasErrors = true
-	}
-
-	for key, value := range capabilities {
-		// Skip informational sibling fields (e.g. dark_mode_detection_detail);
-		// see detailSuffix for the contract.
-		if strings.HasSuffix(key, detailSuffix) {
-			continue
-		}
-
-		status, _ := value.(string)
-
-		components["capability."+key] = status
-		if key != "platform" && !capabilityStatusSupported(status) {
-			hasErrors = true
-		}
-	}
-	// Service health checks (accessibility + overlay per service)
-	serviceChecks := map[string]map[string]error{}
-	if h.hintService != nil {
-		serviceChecks["hints"] = h.hintService.Health(ctx)
-	} else {
-		components["hints"] = healthNotInitialized
-		hasErrors = true
-	}
-
-	if h.gridService != nil {
-		serviceChecks["grid"] = h.gridService.Health(ctx)
-	} else {
-		components["grid"] = healthNotInitialized
-		hasErrors = true
-	}
-
-	if h.actionService != nil {
-		serviceChecks["action"] = h.actionService.Health(ctx)
-	} else {
-		components["action"] = healthNotInitialized
-		hasErrors = true
-	}
-
-	if h.scrollService != nil {
-		serviceChecks["scroll"] = h.scrollService.Health(ctx)
-	} else {
-		components["scroll"] = healthNotInitialized
-		hasErrors = true
-	}
-	// Flatten service sub-checks into components map
-	for service, checks := range serviceChecks {
-		for check, err := range checks {
-			key := service + "." + check
-
-			if err != nil {
-				components[key] = err.Error()
-				hasErrors = true
-
-				h.logger.Warn("Health check failed",
-					zap.String("service", service),
-					zap.String("check", check),
-					zap.Error(err))
-			} else {
-				components[key] = "ok"
-			}
-		}
-	}
-
-	// --- metadata -----------------------------------------------------------
-	configPath := h.ResolveConfigPath()
-
-	mode := ""
-	if h.appState != nil {
-		mode = domain.ModeString(h.appState.CurrentMode())
-	}
-
-	data := map[string]any{
-		"version":      ipc.BuildVersion(),
-		"config":       configPath,
-		"mode":         mode,
-		"capabilities": capabilities,
-		"profile":      profileMap(platform.CurrentProfile()),
-		"components":   components,
-	}
-
-	response := ipc.Response{
-		Success: !hasErrors,
-		Data:    data,
-		Code:    ipc.CodeOK,
-	}
-
-	if hasErrors {
-		response.Code = ipc.CodeActionFailed
-	}
-
-	return response
-}
-
-func (h *InfoHandler) handleConfigSet(ctx context.Context, cmd ipc.Command) ipc.Response {
-	if len(cmd.Args) < minConfigSetArgs {
-		return ipc.Response{
-			Success: false,
-			Message: "config-set requires key and value arguments",
-			Code:    ipc.CodeInvalidInput,
-		}
-	}
-
-	key := cmd.Args[0]
-	value := cmd.Args[1]
-	noReload := len(cmd.Args) > minConfigSetArgs && cmd.Args[minConfigSetArgs] == "--no-reload"
-
-	if noReload {
-		// Lightweight path: update in-memory config and persist without
-		// disrupting active hotkeys or exiting the current mode. Useful
-		// when setting multiple fields in sequence. Caller should run
-		// "neru config reload" afterward to apply all changes.
-		return h.handleConfigSetNoReload(ctx, key, value)
-	}
-
-	// Full path: delegate to the app-level callback for reconfiguration
-	// (component updates, hotkey re-registration, etc.).
-	if h.setConfigField != nil {
-		err := h.setConfigField(ctx, key, value)
-		if err != nil {
-			return ipc.Response{
-				Success: false,
-				Message: err.Error(),
-				Code:    ipc.CodeActionFailed,
-			}
-		}
-
-		return ipc.Response{
-			Success: true,
-			Message: fmt.Sprintf("config %s set to %q", key, value),
-			Code:    ipc.CodeOK,
-		}
-	}
-
-	// Fallback (e.g. tests with no callback): update in-memory config only.
-	return h.handleConfigSetInMemory(ctx, key, value)
-}
-
-// handleConfigSetNoReload updates the in-memory config and persists to the
-// override file without triggering mode exit or hotkey re-registration.
-func (h *InfoHandler) handleConfigSetNoReload(
-	_ context.Context,
-	key, value string,
-) ipc.Response {
-	cfg := h.configSnapshot()
-	if cfg == nil {
-		return h.configNotAvailableResponse()
-	}
-
-	newCfg, err := config.DeepCopyConfig(cfg)
-	if err != nil {
-		return ipc.Response{
-			Success: false,
-			Message: "failed to copy config: " + err.Error(),
-			Code:    ipc.CodeActionFailed,
-		}
-	}
-
-	setErr := config.SetField(newCfg, key, value)
-	if setErr != nil {
-		return ipc.Response{
-			Success: false,
-			Message: setErr.Error(),
-			Code:    ipc.CodeInvalidInput,
-		}
-	}
-
-	// Skip Validate() here so interdependent fields (e.g. grid_cols + keys)
-	// can be updated incrementally before a final "neru config reload".
-	h.configService.Replace(newCfg)
-	h.configMu.Lock()
-	h.config = newCfg
-	h.configMu.Unlock()
-
-	// Persist to override file so changes survive restart.
-	persistErr := h.configService.SaveOverrideField(key, value)
-	if persistErr != nil {
-		return ipc.Response{
-			Success: false,
-			Message: "config set but failed to persist: " + persistErr.Error(),
-			Code:    ipc.CodeActionFailed,
-		}
-	}
-
-	return ipc.Response{
-		Success: true,
-		Message: fmt.Sprintf("config %s set to %q (reload required to apply)", key, value),
-		Code:    ipc.CodeOK,
-	}
-}
-
-// handleConfigReset removes a field from the override file, reverting it to
-// the base config or default value on the next reload.  When --no-reload is
-// present the in-memory config is left alone so callers can batch multiple
-// resets before a final "neru config reload".
-func (h *InfoHandler) handleConfigReset(ctx context.Context, cmd ipc.Command) ipc.Response {
-	if len(cmd.Args) < 1 {
-		return ipc.Response{
-			Success: false,
-			Message: "config-reset requires a key argument",
-			Code:    ipc.CodeInvalidInput,
-		}
-	}
-
-	key := cmd.Args[0]
-	noReload := len(cmd.Args) > 1 && cmd.Args[1] == "--no-reload"
-
-	if config.ConfigFieldType(key) == "unknown" {
-		return ipc.Response{
-			Success: false,
-			Message: "unknown config field: " + key,
-			Code:    ipc.CodeInvalidInput,
-		}
-	}
-
-	removeErr := h.configService.RemoveOverrideField(key)
-	if removeErr != nil {
-		return ipc.Response{
-			Success: false,
-			Message: "failed to remove override: " + removeErr.Error(),
-			Code:    ipc.CodeActionFailed,
-		}
-	}
-
-	if noReload {
-		return ipc.Response{
-			Success: true,
-			Message: fmt.Sprintf("config %s reset (reload required to apply)", key),
-			Code:    ipc.CodeOK,
-		}
-	}
-
-	// Full reload to apply the reset immediately.
-	return h.handleReloadConfig(ctx, ipc.Command{})
-}
-
-// handleConfigSetInMemory updates the in-memory config without persisting.
-// This is the fallback path used when no app-level callback is registered
-// (e.g. in tests).
-func (h *InfoHandler) handleConfigSetInMemory(
-	_ context.Context,
-	key, value string,
-) ipc.Response {
-	cfg := h.configSnapshot()
-	if cfg == nil {
-		return h.configNotAvailableResponse()
-	}
-
-	newCfg, err := config.DeepCopyConfig(cfg)
-	if err != nil {
-		h.logger.Error("Failed to deep copy config", zap.Error(err))
-
-		return ipc.Response{
-			Success: false,
-			Message: "failed to copy config: " + err.Error(),
-			Code:    ipc.CodeActionFailed,
-		}
-	}
-
-	setErr := config.SetField(newCfg, key, value)
-	if setErr != nil {
-		return ipc.Response{
-			Success: false,
-			Message: setErr.Error(),
-			Code:    ipc.CodeInvalidInput,
-		}
-	}
-
-	updateErr := h.configService.Update(newCfg)
-	if updateErr != nil {
-		h.logger.Error("Failed to update config service", zap.Error(updateErr))
-
-		return ipc.Response{
-			Success: false,
-			Message: "failed to apply config: " + updateErr.Error(),
-			Code:    ipc.CodeActionFailed,
-		}
-	}
-
-	h.configMu.Lock()
-	h.config = newCfg
-	h.configMu.Unlock()
-
-	return ipc.Response{
-		Success: true,
-		Message: fmt.Sprintf("config %s set to %q", key, value),
-		Code:    ipc.CodeOK,
-	}
-}
-
-// configNotAvailableResponse returns a standardized response for when the
-// config is nil. Extracted as a helper to avoid repeating the string literal.
-func (h *InfoHandler) configNotAvailableResponse() ipc.Response {
-	return ipc.Response{
-		Success: false,
-		Message: "config not available",
-		Code:    ipc.CodeActionFailed,
-	}
-}
-
-func (h *InfoHandler) systemCapabilities() ports.PlatformCapabilities {
-	if h.systemPort != nil {
-		return h.systemPort.Capabilities()
-	}
-
-	return ports.PlatformCapabilities{}
-}
-
-func capabilitiesMap(capabilities ports.PlatformCapabilities) map[string]any {
-	if capabilities.Platform == "" {
-		return map[string]any{}
-	}
-
-	out := map[string]any{"platform": capabilities.Platform}
-
-	for _, entry := range capabilities.Entries() {
-		out[string(entry.Key)] = capabilityString(entry.FeatureCapability)
-	}
-
-	// Surface dark-mode Detail as a sibling field so `neru doctor` can render
-	// the current live state (e.g. "current state: dark (source=xdg-portal)")
-	// without having to expand FeatureCapability into its own structured map.
-	// Only Linux probes a live state into Detail; Darwin/Windows leave a static
-	// description there, so gate on platform to keep it out of their output.
-	if detail := capabilities.DarkModeDetection.Detail; detail != "" &&
-		strings.HasPrefix(capabilities.Platform, "linux") {
-		out[string(ports.CapabilityDarkModeDetection)+detailSuffix] = detail
-	}
-
-	return out
-}
-
-func profileMap(profile platform.Profile) map[string]any {
-	return map[string]any{
-		"os":                          string(profile.OS),
-		"primary_modifier":            profile.PrimaryModifier,
-		"display_server":              string(profile.DisplayServer),
-		"accessibility_backend":       profile.Accessibility.Name,
-		"accessibility_build_mode":    string(profile.Accessibility.BuildMode),
-		"accessibility_notes":         profile.Accessibility.Notes,
-		"hotkeys_backend":             profile.Hotkeys.Name,
-		"hotkeys_build_mode":          string(profile.Hotkeys.BuildMode),
-		"hotkeys_notes":               profile.Hotkeys.Notes,
-		"keyboard_capture_backend":    profile.KeyboardCapture.Name,
-		"keyboard_capture_build_mode": string(profile.KeyboardCapture.BuildMode),
-		"keyboard_capture_notes":      profile.KeyboardCapture.Notes,
-		"overlay_backend":             profile.Overlay.Name,
-		"overlay_build_mode":          string(profile.Overlay.BuildMode),
-		"overlay_notes":               profile.Overlay.Notes,
-		"notifications_backend":       profile.Notifications.Name,
-		"notifications_build_mode":    string(profile.Notifications.BuildMode),
-		"notifications_notes":         profile.Notifications.Notes,
-	}
-}
-
-func capabilityString(capability ports.FeatureCapability) string {
-	if capability.Status == "" {
-		return string(ports.FeatureStatusStub)
-	}
-
-	return string(capability.Status)
-}
-
-func capabilityStatusSupported(status string) bool {
-	capability := ports.FeatureCapability{Status: ports.FeatureStatus(status)}
-
-	return capability.Supported()
 }
